Update component names and modify api call for coordinates
