+++ .env
@@ -1,5 +1,5 @@
-HIKING_PROJECT_ALL_TRAILS_URL=https://www.hikingproject.com/data/get-trails?
-HIKING_PROJECT_KEY=200671559-f2a171f9f153559a7b741d306678425a
-COORDINATES_KEY=eJwz5DQ0AAFLM1NDzmoDU3NzF2MjR11nAzdHXUdnN3NdYwtHR103U3NTM3MXR2NXR-daABOdC2A
-COORDINATES_UNIQUE_USER=&user=eJwz5DQ0AAFLM1NDzmoDU3NzF2MjR11nAzdHXUdnN3NdYwtHR103U3NTM3MXR2NXR-daABOdC2A
-COORDINATES_URL=https://global.skyhookwireless.com/wps2/json/location?key=
+REACT_APP_HIKING_PROJECT_ALL_TRAILS_URL=https://www.hikingproject.com/data/get-trails?
+REACT_APP_HIKING_PROJECT_KEY=200671559-f2a171f9f153559a7b741d306678425a
+REACT_APP_COORDINATES_URL=https://global.skyhookwireless.com/wps2/json/location?
+REACT_APP_COORDINATES_KEY=key=eJwz5DQ0AAFLM1NDzmoDU3NzF2MjR11nAzdHXUdnN3NdYwtHR103U3NTM3MXR2NXR-daABOdC2A
+REACT_APP_COORDINATES_UNIQUE_USER=&user=eJwz5DQ0AAFLM1NDzmoDU3NzF2MjR11nAzdHXUdnN3NdYwtHR103U3NTM3MXR2NXR-daABOdC2A
