+++ src/Components/App/App.js
@@ -1,10 +1,11 @@
 import React from 'react';
+import Trails from '../Trails/Trails';
 import './App.css';
 
 function App() {
   return (
     <div className="App">
-
+      <Trails />
     </div>
   );
 }
